+++ core-aam/aamtests/conftest.py
@@ -7,6 +7,7 @@ sys.path.insert(0, str(webdriver_tests_path))
 
 pytest_plugins = (
     "pytest_fixtures.webdriver",
+    "pytest_fixtures.webdriver_classic",
     "tests.support.fixtures",
     "tests.support.classic.fixtures",
     "aamtests.support.fixtures_a11y_api",
