+++ tools/requirements_pytest.txt
@@ -1,3 +1,4 @@
 pytest==8.3.2
 pytest-cov==7.1.0
+pytest-asyncio==1.2.0
 hypothesis==6.112.0
