+++ webdriver/tests/conftest.py
@@ -1,5 +1,6 @@
 pytest_plugins = (
     "pytest_fixtures.webdriver",
+    "pytest_fixtures.webdriver_classic",
     "tests.support.fixtures",
     "tests.support.bidi.fixtures",
     "tests.support.classic.fixtures",
